Gui: Split declaration and definition of ToolBarAreaWidget
In order to support targeting ToolBarAreaWidgets via the QSS we need to
make it a proper QObject using Q_OBJECT macro, which must be placed in
header files.
